// src/app/_dto/tblwhmdetail.dto.ts
export class TblWhMDetailDto {

    StoS_SysID: number = 0;
    StoSS_SysID: number = 0;

    StoS_Prod_SysID: number = 0;
    StoS_Prod_Code: string = null;
    StoS_Prod_Name: string = null;
    StoS_Prod_LName: string = null;

    StoS_Rak_SysID: number = 0;
    StoS_Rak_Code: string = null;
    StoS_Rak_Name: string = null;

    StoS_Batch_No: string = null;
    StoS_Store_Qty: number = 0;

    StoS_PalletDiam_SysID: number = 0;
    StoS_PalletDiam_Code: string = null;
    StoS_PalletDiam_Name: string = null;
    StoS_PalletDiam_Length: string = null;
    StoS_PalletDiam_Width: string = null;
    StoS_PalletDiam_Hight: string = null;
    StoS_PalletGross_Weight: number = 0;
    StoS_PalletNet_Weight: number = 0;

    StoS_ItemtDiam_SysID: number = 0;
    StoS_ItemtDiam_Code: string = null;
    StoS_ItemDiam_Name: string = null;
    StoS_Item_Unit: string = null;
    StoS_Item_Length: string = null;
    StoS_Item_Width: string = null;
    StoS_Item_Hight: string = null;

    StoS_Rem_SysID: number = 0;
    StoS_Rem_Code: string = null;
    StoS_Rem_Name: string = null;
}